backend/requirements-railway.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
jinja2==3.1.2
openpyxl==3.1.2
reportlab==4.0.7
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
